remove company names
